requirements.txt
torch
torchvision
torchaudio
transformers
gradio
pillow
sentencepiece
sympy
mpmath
sacremoses
